fix(review): MEDIUM/LOW remediation pass — Bands 1 & 2
First two bands of the deferred full-review backlog: trivial one-liners plus
easy localized fixes. Verified by dotnet build/test (Memory 26, App 353) and
Vitest (110). Closes the MEM-4..10/12 items deferred in e86e2f3.

Memory:   MEM-4/6/7/8/9/10/12 — GetKey short-key guard, volatile exception
          flag, dead GetProcessName removed, tokenizer min/packed-length
          guards, (int) PID cast, payload chained into Exception.Data,
          InvariantCulture hex parse (+3 regression tests)
Config:   CFG-7/8/9 — active-profile check under lock, per-call StringBuilder,
          only-succeeded deletions cleared
WebView:  WVH-5/6/7/9, SIF-3/4 — TickCount64 throttle, env-task null guard,
          console-forward cap, PreserveSig GetFrameStatistics, faulted-task
          and EndInvoke logging
Bridge:   BRG-12/14 — typed Task<T> unwrap, correlation-id wrap
Chat/TS:  CHT-6, TSO-10/11, TSS-11/12 — read-only Marks, empty-source guard,
          audio.play rejection, console.error, opener null guard
LogConv:  UPD-10/11/12/13 — progress division, regex suffix append, TryParse
          skip, duplicate guard removed

UPD-8/9 were already closed in af4c894; the audit at 928f552 pre-dated that.

+++ src/Gobchat.LogConverter/LogConverter.cs
@@ -81,7 +81,7 @@ namespace Gobchat.LogConverter
 
             for (var lineIdx = 0; lineIdx < lines.Length; ++lineIdx)
             {
-                _monitor.Progress = lineIdx / lines.Length;
+                _monitor.Progress = (double)lineIdx / lines.Length;
                 var line = lines[lineIdx];
 
                 try
